fix(persistence): compare mixed string/number ids by numeric value
String coercion only matched decimals whose string form was canonical, so
a scaled decimal key hydrated as "70.000" did not match the in-memory 70.
Compare the mixed string/number single-id fast path by numeric value
instead, so bigints returned as "1" and scaled decimals both match, while
still rejecting non-finite values and unsafe integers beyond
MAX_SAFE_INTEGER where a JS number cannot represent the id exactly.

Give the #11773 fixture an int parent PK and a bigint child PK so the
regression isolates the child orphan/FK-null path (a bigint parent could
otherwise fail parent matching first).

+++ packages/typeorm/test/github-issues/11773/entity/Child.ts
@@ -9,10 +9,12 @@ import { Parent } from "./Parent"
 
 @Entity()
 export class Child {
+    // The child's own PK is the bigint that some drivers hydrate as a string.
     @PrimaryColumn({ type: "bigint" })
     id: number
 
-    @Column({ type: "bigint", nullable: true })
+    // FK matches the parent's int PK; it is the column that was nulled.
+    @Column({ type: "int", nullable: true })
     parentId: number | null
 
     @ManyToOne(() => Parent, (parent) => parent.children)
