+++ packages/typeorm/test/github-issues/11773/entity/Parent.ts
@@ -3,7 +3,9 @@ import { Child } from "./Child"
 
 @Entity()
 export class Parent {
-    @PrimaryColumn({ type: "bigint" })
+    // Parent uses an int PK so it always matches on re-save; only the child's
+    // bigint PK exercises the string/number mismatch that nulled the FK.
+    @PrimaryColumn({ type: "int" })
     id: number
 
     @OneToMany(() => Child, (child) => child.parent, { cascade: true })
